行业、会计准则获取 key 调整

+++ src/main/java/com/github/outerman/be/engine/util/CommonUtil.java
@@ -121,7 +121,7 @@ public final class CommonUtil {
         return name;
     }
 
-    public static String getAccountingStandardName(Integer id) {
+    public static String getAccountingStandardName(Long id) {
         String name;
         if (id == 18) {
             name = "企业会计准则2007";
